Casted the amount from onTip to be a number
The amount variable being returned by the onTip() function was being returned as a string as this is how it is parsed from DogeChat. For simplicity I cast this to a number before adding it to the object of data as it seems appropriate the amount should be received as a number rather than a string.